Drop (both) includes of deprecated libmesh mesh_inserter_iterator.h header
Also remove a reference in a comment to the deprecated
mesh_inserter_iterator class.  MOOSE still defines and uses the
extra_ghost_elem_inserter, but it is not derived from
mesh_inserter_iterator in libmesh, so there should be no issue with
dropping this include.

Part of an ongoing deprecated code cleanup, (see also: #31732).