Catch possible "det != 0" exceptions from libMesh
We only had one test failing from this, but in theory any run that
inverts an element side and tries to reinit on that side could be
failing here.

I'd be tempted to refactor this code at this point, but once we get the
next libMesh update in we'll just be catching via proper exception type
and can get rid of the string parsing.